api/requirements.txt
flask==3.1.2
flask-cors==6.0.2
pandas==2.1.1
numpy==1.27.5
scipy==1.11.2
matplotlib==3.8.1
seaborn==0.12.3
gprofiler-official==1.0.1
gunicorn==23.2.0
